Fix `get_api_definitions` to respect descriptors
In Python 3.3, `inspect.getmembers` doesn't respect the descriptor
protocol when the first argument is a class. In other words, the
function includes any descriptors bound to `service` as is rather than
calling the descriptor's __get__ method. This is seemingly fixed in
Python 2.7 and 3.4+. Directly invoking `getattr` to force Python's
attribute lookup protocol is a decent workaround to ensure parity: