Remove documentation for `with_value` method
For the use cases that this method supports, there are more elegant
alternatives available or planned for v0.6.0. For `Header` and
`HeaderMap`, the `headers` method annotation supports static headers and
the new `session` property of `Consumer` allows for dynamic headers.
Similar alternatives are being introduced in v0.6.0 for `Query` and
`QueryMap`