Add and expose auth methods: MultiAuth ApiToken*
This exposes all auth methods and adds these additional auth classes:
 - MultiAuth
 - ApiTokenParam
 - ApiTokenHeader

The purpose of MultiAuth is to allow for multiple auth methods. This may
be required when an API user needs to supply intermediate credentials--
as when using ProxyAuth to authenticate with an intermediate proxy--
or when an API requires multiple auth tokens. MultiAuth acts like a list
so that users can append, extend or iterate over the contained auth
methods.

The purpose of the the ApiToken* auth methods is to allow both direct
API users and API library authors to easily provide auth tokens in a
standardized manner.

When directly supplying the auth token, some users may prefer to supply
both the parameter/header name and the token at the same time. Thus, we
allow such users to directly use the generic ApiToken* auth methods.

For API library authors that want to supply the parameter/header name in
a reusable way (so users only have to supply the token), they can
use properties on an ApiToken* subclass to define the relvant query
parameter or header.

This also refactors BasicAuth and ProxyAuth as subclasses of
ApiTokenHeader to harmonize the implementation details of all the auth
methods. Any uplink users that directly accessed the private _auth_str
attribute will need to replace that with _header_value.